Removed mock prefixes and renamed to mock
